Class1/pyproject.toml
[project]
name = "class1"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents>=1.4",
    "livekit-plugins-cartesia>=1.4.4",
    "livekit-plugins-noise-cancellation>=0.2.5",
    "livekit-plugins-openai>=1.4.4",
    "livekit-plugins-silero>=1.4.4",
    "livekit-plugins-turn-detector>=1.4.4",
    "python-dotenv>=1.2.2",
]
